Remove dependency on python-cliff-tablib
python-cliff now contributes its own (and better) json and yaml
formatters so python-cliff-tablib is now not required.

If it is installed then it contributes duplicate json and yaml
formatters. Since it is considered dead upstream then
the python-cliff-tablib package should be modified to not install
anything, or be obsoleted by python-cliff

Change-Id: I29856c7dc08884dc0f58f15f685e94197f97e347