Create a single major-versioned binary instead of two ones
Currently, we are following old fedora policy to create binaries

 - foo
 - foo-<major>, i.e.: foo-3
 - foo-<minor>, i.e. foo-3.6

As we are moving to single package, having versioned binaries is
not longer required. However, for backwards compatibility i think
we must keep the major version one, which in some cases has been
used, for example in spec files. I'd say the minor-version one
is not longer needed and add some extra complexity as adding a
new macro or reusing python_version one which is problematic
because there are two different implementations with different
result in both standard python macros and openstack ones.

Change-Id: I958262cb46911776edd760555a177c1a34e3fd52